Finished image uploading frontend. Needs image delete button

+++ app/api/routes/containerimages.php
@@ -10,7 +10,9 @@ class containerimagesRoute extends rest
         return $this->error("Invalid container");
       }
       $ci = Containerimage::findOne(array("containerid" => intval($this->id)));
-      
+      if (!$ci) {
+        return $this->error("Image not found", self::HTTP_NOT_FOUND);
+      }
       switch(strtolower(substr($ci->filename, strrpos($ci->filename, ".")+1))) {
           case "gif":
               $imagetype = IMAGETYPE_GIF;
@@ -34,12 +36,16 @@ class containerimagesRoute extends rest
     if (!$c) {
       return $this->error("Invalid container");
     }
+    $ci = Containerimage::findOne(array("containerid" => intval($this->id)));
+    if ($ci) {
+      $ci->delete();
+    }
     $f = new Containerimage();
     $f->filename = $_FILES['file']['name'];
-    if (!$f->blah($_FILES['file']['tmp_name'])) {
+    if (!$f->uploadFromLocation($_FILES['file']['tmp_name'])) {
         return $this->error("Invalid file");
     }
-    $f->itemid = intval($this->id);
+    $f->containerid = intval($this->id);
     $f->save();
   }
   
@@ -49,6 +55,9 @@ class containerimagesRoute extends rest
       return $this->error("Invalid container");
     }
     $ci = Containerimage::findOne(array("containerid" => intval($this->id)));
+    if (!$ci) {
+      return $this->error("Image not found", self::HTTP_NOT_FOUND);
+    }
     $ci->delete();
   }
 }
